requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
motor==3.1.2
pymongo==4.3.3
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv
email-validator>=1.3.0
python-multipart
python-jose[cryptography]
authlib
httpx
app
itsdangerous
aiomysql